+++ Assets/Plugins/GameService/Utils/RealTimeUtil/Classes/Observers/RealtimeRigidbodyObserver.cs
@@ -66,6 +66,8 @@ namespace Plugins.GameService.Utils.RealTimeUtil.Classes.Observers
         
         public void FixedUpdate()
         {
+            if(GsLiveRealtime.IsCurrentPlayerObserving(this)) return;
+
             _rBody.position = Vector3.MoveTowards(_rBody.position, _mNetworkPosition, _mDistance * (1.0f / GsLiveRealtime.SerializationRate));
             _rBody.rotation = Quaternion.RotateTowards(_rBody.rotation, _mNetworkRotation, _mAngle * (1.0f / GsLiveRealtime.SerializationRate));
         }