+++ Assets/Plugins/GameService/Utils/RealTimeUtil/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 
 
+## [1.6.0] Stable - 2021-06-28
+
+### Added
+
+- Added Check CurrentPlayerObserving To Avoid Update In Observer Player
+
 
 ## [1.5.1] Stable - 2021-06-26